read delta schema
